condition_verifier/requirements.txt
flask
opencv-python
numpy
tensorflow  # or torch, depending on AI models used
requests
pyyaml

# External verifier program dependency (if required)
custom-verifier